mlflow-mlops-workshop/requirements.txt
mlflow==2.10.2
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
jupyter==1.0.0
notebook==7.0.3

tensorflow==2.15.0
keras==2.15.0

torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
pytorch-lightning==2.1.3

gymnasium==0.29.1
stable-baselines3==2.2.1

optuna==3.5.0
hyperopt==0.2.7
xgboost==2.0.3
lightgbm==4.1.0
catboost==1.2.2

pyspark==3.5.0

apache-airflow==2.8.1
apache-airflow-providers-docker==3.8.0
croniter==2.0.1

pillow==10.1.0
opencv-python==4.8.1.78
plotly==5.18.0

flask==3.0.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

pytest==7.4.3
black==23.12.1
flake8==7.0.0

psutil==5.9.6
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2
